// packages/drool-sdk-ext/src/protocol/session/sources/enums.ts
export enum SessionPlatform {
  Slack = 'slack',
  Web = 'web',
  Api = 'api',
  SessionsApi = 'sessions_api',
  Linear = 'linear',
  ReadinessRemediation = 'readiness-remediation',
  ReadinessEvaluation = 'readiness-evaluation',
  Automation = 'automation',
  WikiGeneration = 'wiki-generation',
  WikiCISetup = 'wiki-ci-setup',
}

export enum SessionOrigin {
  Web = 'web',
  Desktop = 'desktop',
  CliTui = 'cli_tui',
  CliExec = 'cli_exec',
  CliAcp = 'cli_acp',
  Slack = 'slack',
  Linear = 'linear',
  SessionsApi = 'sessions_api',
  Api = 'api',
  Automation = 'automation',
  ReadinessRemediation = 'readiness-remediation',
  ReadinessEvaluation = 'readiness-evaluation',
  WikiGeneration = 'wiki-generation',
  WikiCiSetup = 'wiki-ci-setup',
}